WIP: Fix events/subscriptions
