+++ src/lib/kotlin/slatekit-jobs/src/main/kotlin/slatekit/jobs/WorkerContext.kt
@@ -3,4 +3,4 @@ package slatekit.jobs
 import slatekit.common.Identity
 import slatekit.common.metrics.Recorder
 
-data class WorkerContext(val jobId: Identity, val worker: Worker<*>, val stats:Recorder<Task, WorkResult>)
+data class WorkerContext(val id: Identity, val worker: Worker<*>, val stats:Recorder<Task, WorkResult>)
