+++ src/lib/kotlin/slatekit-jobs/src/main/kotlin/slatekit/jobs/events/JobEvents.kt
@@ -1,0 +1,10 @@
+package slatekit.jobs.events
+
+import slatekit.jobs.Job
+
+class JobEvents : SubscribedEvents<Job>(){
+
+    override suspend fun notify(item: Job) {
+        notify(item, item.status())
+    }
+}
