+++ src/lib/kotlin/slatekit-jobs/src/main/kotlin/slatekit/jobs/events/WorkerEvents.kt
@@ -1,0 +1,24 @@
+package slatekit.jobs.events
+
+import slatekit.functions.policy.Policy
+import slatekit.jobs.WorkRequest
+import slatekit.jobs.WorkResult
+import slatekit.jobs.Worker
+import slatekit.jobs.Workers
+
+class WorkerEvents(val workers: Workers) : SubscribedEvents<Worker<*>>(), WorkEvents {
+    val policies = mutableListOf<Policy<WorkRequest, WorkResult>>()
+
+    override suspend fun notify(item: Worker<*>) {
+        notify(item, item.status())
+    }
+
+
+    /**
+     * Applies the policy to these workers
+     * Policies add behaviour to workers such as retries, limits, error ratios, etc
+     */
+    override fun apply(policy: Policy<WorkRequest, WorkResult>) {
+        policies.add(policy)
+    }
+}
